Split functions into logically named files
